+++ builders.json
@@ -1,15 +1,14 @@
 {
   "builders": {
     "cordova-build": {
-      "implementation": "./builders/cordova-build/index2",
-      "class": "./builders/cordova-build",
+      "implementation": "./builders/cordova-build/index",
       "schema": "./builders/cordova-build/schema.json",
       "description": "Perform a browser build with Cordova assets."
     },
 
 
     "cordova-serve": {
-      "implementation": "./builders/cordova-serve/index2",
+      "implementation": "./builders/cordova-serve/index",
       "schema": "./builders/cordova-serve/schema.json",
       "description": "Run the dev-server with Cordova assets."
     }
