+++ builders/cordova-serve/index.ts
@@ -1,82 +1,43 @@
-import { BuildEvent, Builder, BuilderConfiguration, BuilderContext, BuilderDescription } from '@angular-devkit/architect';
-import { NormalizedBrowserBuilderSchema } from '@angular-devkit/build-angular/src/browser/schema';
-import { DevServerBuilder, DevServerBuilderOptions } from '@angular-devkit/build-angular/src/dev-server';
-import { Path, virtualFs } from '@angular-devkit/core';
-import * as fs from 'fs';
-import { Observable, from, of } from 'rxjs';
-import { concatMap, tap } from 'rxjs/operators';
+import { BuilderContext, BuilderOutput, createBuilder, targetFromTargetString } from '@angular-devkit/architect';
+import { DevServerBuildOutput } from '@angular-devkit/build-webpack';
+import { json } from '@angular-devkit/core';
 
-import { CordovaBuildBuilder, CordovaBuildBuilderSchema } from '../cordova-build';
+import { prepareBrowserConfig } from '../utils';
 
 import { createConsoleLogServer } from './log-server';
 import { CordovaServeBuilderSchema } from './schema';
 
-export class CordovaServeBuilder implements Builder<CordovaServeBuilderSchema> {
-  constructor(public context: BuilderContext) {}
+export type CordovaDevServerBuilderOptions = CordovaServeBuilderSchema & json.JsonObject;
 
-  run(builderConfig: BuilderConfiguration<CordovaServeBuilderSchema>): Observable<BuildEvent> {
-    const { options: cordovaServeOptions } = builderConfig;
-    const { devServerTarget, port, host, ssl } = cordovaServeOptions;
-    const [ project, target, configuration ] = devServerTarget.split(':');
+export async function serveCordova(
+  options: CordovaServeBuilderSchema,
+  context: BuilderContext
+): Promise<BuilderOutput> {
+  return new Promise(async () => {
+    context.reportStatus(`running cordova serve...`);
+    const { devServerTarget, cordovaBuildTarget, port, host, ssl } = options;
 
-    const devServerTargetSpec = { project, target, configuration, overrides: { port, host, ssl } };
-    const devServerBuilderConfig = this.context.architect.getBuilderConfiguration<DevServerBuilderOptions>(devServerTargetSpec);
+    // Getting the original browser build options
+    const cordovaBuildTargetSpec = targetFromTargetString(cordovaBuildTarget);
+    const cordovaBuildTargetOptions = await context.getTargetOptions(cordovaBuildTargetSpec) as { browserTarget: string };
+    const browserBuildTargetSpec = targetFromTargetString(cordovaBuildTargetOptions.browserTarget);
 
-    let devServerDescription: BuilderDescription;
-    let cordovaBuildConfig: BuilderConfiguration<CordovaBuildBuilderSchema>;
+    // What we actually need....
+    const browserBuildTargetOptions = await context.getTargetOptions(browserBuildTargetSpec);
 
-    return of(null).pipe(
-      concatMap(() => this.context.architect.getBuilderDescription(devServerBuilderConfig)),
-      tap(description => devServerDescription = description),
-      concatMap(() => this.context.architect.validateBuilderOptions(devServerBuilderConfig, devServerDescription)),
-      concatMap(() => this._getCordovaBuildConfig(cordovaServeOptions)),
-      tap(config => cordovaBuildConfig = config),
-      concatMap(() => of(new CordovaDevServerBuilder(this.context, cordovaBuildConfig.options))),
-      concatMap(builder => builder.run(devServerBuilderConfig))
-    );
-  }
+    // Modifying those options to pass in cordova-speicfic stuff
+    prepareBrowserConfig(options, browserBuildTargetOptions);
 
-  protected _getCordovaBuildConfig(cordovaServeOptions: CordovaServeBuilderSchema): Observable<BuilderConfiguration<CordovaBuildBuilderSchema>> {
-    const {
-      platform,
-      cordovaBasePath,
-      cordovaAssets,
-      cordovaMock,
-      consolelogs,
-      consolelogsPort,
-      sourceMap,
-    } = cordovaServeOptions;
-
-    const [ project, target, configuration ] = cordovaServeOptions.cordovaBuildTarget.split(':');
-    const cordovaBuildTargetSpec = { project, target, configuration, overrides: { platform, cordovaBasePath, cordovaAssets, cordovaMock, consolelogs, consolelogsPort, sourceMap } };
-    const cordovaBuildTargetConfig = this.context.architect.getBuilderConfiguration<CordovaBuildBuilderSchema>(cordovaBuildTargetSpec);
-
-    return this.context.architect.getBuilderDescription(cordovaBuildTargetConfig).pipe(
-      concatMap(cordovaBuildDescription => this.context.architect.validateBuilderOptions(cordovaBuildTargetConfig, cordovaBuildDescription))
-    );
-  }
-}
-
-class CordovaDevServerBuilder extends DevServerBuilder {
-  constructor(context: BuilderContext, public cordovaBuildOptions: CordovaBuildBuilderSchema) {
-    super(context);
-  }
-
-  run(builderConfig: BuilderConfiguration<DevServerBuilderOptions>): Observable<BuildEvent> {
-    if (this.cordovaBuildOptions.consolelogs && this.cordovaBuildOptions.consolelogsPort) {
-      return from(createConsoleLogServer(builderConfig.options.host, this.cordovaBuildOptions.consolelogsPort))
-        .pipe(_ => super.run(builderConfig));
+    if (options.consolelogs && options.consolelogsPort) {
+      await createConsoleLogServer(host, options.consolelogsPort);
     }
-    return super.run(builderConfig);
-  }
 
-  buildWebpackConfig(root: Path, projectRoot: Path, host: virtualFs.Host<fs.Stats>, browserOptions: NormalizedBrowserBuilderSchema) {
-    const builder = new CordovaBuildBuilder(this.context);
-    builder.validateBuilderConfig(this.cordovaBuildOptions);
-    builder.prepareBrowserConfig(this.cordovaBuildOptions, browserOptions);
+    const devServerTargetSpec = targetFromTargetString(devServerTarget);
+    const devServerTargetOptions = await context.getTargetOptions(devServerTargetSpec);
 
-    return super.buildWebpackConfig(root, projectRoot, host, browserOptions);
-  }
+    return context
+      .scheduleTarget(devServerTargetSpec, { host, port, ssl }, devServerTargetOptions)
+      .then(buildEvent => ({ ...buildEvent }));
+  });
 }
-
-export default CordovaServeBuilder;
+export default createBuilder<CordovaDevServerBuilderOptions, DevServerBuildOutput>(serveCordova);
